refactor: moved ModelType to core package
