+++ src/main/java/vaultiq/session/config/annotation/model/VaultiqPersistenceMethod.java
@@ -1,11 +1,13 @@
 package vaultiq.session.config.annotation.model;
 
+import vaultiq.session.core.model.ModelType;
+
 /**
  * Enumeration defining the individual persistence methods that can be enabled
  * for a Vaultiq session data model.
  * <p>
  * This enum represents the specific storage mechanisms (cache or JPA) that can be
- * configured for a given {@link vaultiq.session.cache.model.ModelType}. It is often used
+ * configured for a given {@link ModelType}. It is often used
  * in configuration properties or annotations (like {@link vaultiq.session.config.annotation.ConditionalOnVaultiqModelConfig})
  * to specify which method of persistence is active for a particular model.
  * </p>
@@ -15,7 +17,7 @@ package vaultiq.session.config.annotation.model;
  * </p>
  *
  * @see vaultiq.session.config.annotation.ConditionalOnVaultiqModelConfig Annotation using this enum to condition beans.
- * @see vaultiq.session.cache.model.ModelType Defines the different types of session data models.
+ * @see ModelType Defines the different types of session data models.
  * @see VaultiqPersistenceMode Defines combinations of persistence methods.
  */
 public enum VaultiqPersistenceMethod {
